Final corrections

+++ src/main/java/com/marijana/library1223/controllers/UserController.java
@@ -1,13 +1,16 @@
 package com.marijana.library1223.controllers;
 
-import com.marijana.library1223.configuration.PaginationConfiguration;
+import com.marijana.library1223.helpers.HandleBindingErrors;
+import com.marijana.library1223.helpers.PaginationConfiguration;
 import com.marijana.library1223.dtos.UserDto;
+import com.marijana.library1223.exceptions.AccessDeniedException;
 import com.marijana.library1223.exceptions.BadRequestException;
 import com.marijana.library1223.services.UserService;
 import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.validation.BindingResult;
-import org.springframework.validation.FieldError;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
@@ -31,12 +34,9 @@ public class UserController {
     public ResponseEntity<Object> createUser(@Valid @RequestBody UserDto userDto, BindingResult bindingResult) {
 
         if(bindingResult.hasErrors()) {
-            StringBuilder stringBuilder = new StringBuilder();
-            for (FieldError fieldError : bindingResult.getFieldErrors()) {
-                stringBuilder.append(fieldError.getDefaultMessage());
-                stringBuilder.append("\n");
-            }
-            return ResponseEntity.badRequest().body(stringBuilder.toString());
+
+            return HandleBindingErrors.handleBindingErrors(bindingResult);
+
         } else {
 
             String newUsername = userService.createNewUser(userDto);
@@ -55,9 +55,17 @@ public class UserController {
 
 
     @GetMapping(value = "/{username}")
-    ResponseEntity<UserDto> getUser(@PathVariable("username") String username) {
-        UserDto optionalUser = userService.getUserByUsername(username);
-        return ResponseEntity.ok().body(optionalUser);
+    ResponseEntity<UserDto> getUser(
+            @PathVariable("username") String username,
+            @AuthenticationPrincipal UserDetails userDetails) {
+
+        if (username.equals(userDetails.getUsername())) {
+            UserDto optionalUser = userService.getUserByUsername(username);
+            return ResponseEntity.ok().body(optionalUser);
+
+        } else {
+            throw new AccessDeniedException("You are not authorized to access this user");
+        }
     }
 
 
@@ -81,18 +89,44 @@ public class UserController {
     }
 
 
-
     @PutMapping("/update/{username}")
-    public ResponseEntity<UserDto> updateOneUser(@PathVariable("username") String username, @RequestBody UserDto userDto) {
-        userService.updateUserPassword(username, userDto);
-        return ResponseEntity.noContent().build();
+    public ResponseEntity<UserDto> updateOneUser(
+            @PathVariable("username") String username,
+            @RequestBody UserDto userDto,
+            @AuthenticationPrincipal UserDetails userDetails) {
+
+        if(username.equals(userDetails.getUsername())) {
+
+            userService.updateUserPassword(username, userDto);
+            return ResponseEntity.noContent().build();
+
+        } else {
+
+            throw new AccessDeniedException("You are not authorized to update this user");
+
+        }
+
     }
 
-    @DeleteMapping(value = "/delete/{username}")
-    public ResponseEntity<Object> deleteOneUser(@PathVariable("username") String username) {
-        userService.deleteUser(username);
-        return ResponseEntity.noContent().build();
+
+    @DeleteMapping(value = "/delete/{username}" )
+    public String deleteUser(
+            @PathVariable("username") String username,
+            @AuthenticationPrincipal UserDetails userDetails) {
+
+        if (username.equals(userDetails.getUsername())) {
+
+            userService.deleteUser(username);
+
+            return "User " + userDetails.getUsername() + " has been deleted: " + username;
+
+        } else {
+            throw new AccessDeniedException("You are not authorized to delete this user");
+        }
+
     }
+
+
 
     //-----authorities
